fix: disable serialisation tests when miri is active
We move the `serialize_types_roundtrip` test to put it under the new cfg flag
